code1.0.5/social-master/server/wey_backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.2.0
nanoid==2.0.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
drf-yasg==1.21.7
schedule==1.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
better-profanity==0.7.0
pyyaml==6.0
requests==2.31.0
requests-toolbelt==1.0.0
typing_extensions==4.5.0
urllib3==2.0.7
nanoid==2.0.0